+++ src/Resources/contao/dca/tl_mitgliederverwaltung.php
@@ -10,7 +10,7 @@ $GLOBALS['TL_DCA']['tl_mitgliederverwaltung'] = array
 	'config' => array
 	(
 		'dataContainer'               => 'Table',
-		'ctable'                      => array('tl_mitgliederverwaltung_applications'),
+		'ctable'                      => array('tl_mitgliederverwaltung_applications', 'tl_mitgliederverwaltung_konto'),
 		'enableVersioning'            => true,
 		'sql' => array
 		(
@@ -88,12 +88,25 @@ $GLOBALS['TL_DCA']['tl_mitgliederverwaltung'] = array
 				'href'                => 'table=tl_mitgliederverwaltung_applications',
 				'icon'                => 'bundles/contaomitgliederverwaltung/images/application.png'
 			),
+			'konto' => array
+			(
+				'label'               => &$GLOBALS['TL_LANG']['tl_mitgliederverwaltung']['konto'],
+				'href'                => 'table=tl_mitgliederverwaltung_konto',
+				'icon'                => 'bundles/contaomitgliederverwaltung/images/euro.png'
+			),
 			'toggle' => array
 			(
-				'label'               => &$GLOBALS['TL_LANG']['tl_mitgliederverwaltung']['toggle'],
-				'icon'                => 'visible.gif',
-				'attributes'          => 'onclick="Backend.getScrollOffset();return AjaxRequest.toggleVisibility(this,%s)"',
-				'button_callback'     => array('tl_mitgliederverwaltung', 'toggleIcon')
+				'label'                => &$GLOBALS['TL_LANG']['tl_mitgliederverwaltung']['toggle'],
+				'attributes'           => 'onclick="Backend.getScrollOffset()"',
+				'haste_ajax_operation' => array
+				(
+					'field'            => 'published',
+					'options'          => array
+					(
+						array('value' => '', 'icon' => 'invisible.svg'),
+						array('value' => '1', 'icon' => 'visible.svg'),
+					),
+				),
 			),
 			'show' => array
 			(
@@ -1460,71 +1473,4 @@ class tl_mitgliederverwaltung extends \Backend
 		return $args;
 	}
 
-	/**
-	 * Ändert das Aussehen des Toggle-Buttons.
-	 * @param $row
-	 * @param $href
-	 * @param $label
-	 * @param $title
-	 * @param $icon
-	 * @param $attributes
-	 * @return string
-	 */
-	public function toggleIcon($row, $href, $label, $title, $icon, $attributes)
-	{
-		$this->import('BackendUser', 'User');
-		
-		if (strlen($this->Input->get('tid')))
-		{
-			$this->toggleVisibility($this->Input->get('tid'), ($this->Input->get('state') == 0));
-			$this->redirect($this->getReferer());
-		}
-		
-		// Check permissions AFTER checking the tid, so hacking attempts are logged
-		if (!$this->User->isAdmin && !$this->User->hasAccess('tl_mitgliederverwaltung::published', 'alexf'))
-		{
-			return '';
-		}
-		
-		$href .= '&amp;id='.$this->Input->get('id').'&amp;tid='.$row['id'].'&amp;state='.$row[''];
-		
-		if (!$row['published'])
-		{
-			$icon = 'invisible.gif';
-		}
-		
-		return '<a href="'.$this->addToUrl($href).'" title="'.specialchars($title).'"'.$attributes.'>'.$this->generateImage($icon, $label).'</a> ';
-	}
-
-	/**
-	 * Toggle the visibility of an element
-	 * @param integer
-	 * @param boolean
-	 */
-	public function toggleVisibility($intId, $blnPublished)
-	{
-		// Check permissions to publish
-		if (!$this->User->isAdmin && !$this->User->hasAccess('tl_mitgliederverwaltung::published', 'alexf'))
-		{
-			$this->log('Not enough permissions to show/hide record ID "'.$intId.'"', 'tl_mitgliederverwaltung toggleVisibility', TL_ERROR);
-			$this->redirect('contao/main.php?act=error');
-		}
-		
-		$this->createInitialVersion('tl_mitgliederverwaltung', $intId);
-		
-		// Trigger the save_callback
-		if (is_array($GLOBALS['TL_DCA']['tl_mitgliederverwaltung']['fields']['published']['save_callback']))
-		{
-			foreach ($GLOBALS['TL_DCA']['tl_mitgliederverwaltung']['fields']['published']['save_callback'] as $callback)
-			{
-				$this->import($callback[0]);
-				$blnPublished = $this->$callback[0]->$callback[1]($blnPublished, $this);
-			}
-		}
-		
-		// Update the database
-		$this->Database->prepare("UPDATE tl_mitgliederverwaltung SET tstamp=". time() .", published='" . ($blnPublished ? '' : '1') . "' WHERE id=?")
-		               ->execute($intId);
-		$this->createNewVersion('tl_mitgliederverwaltung', $intId);
-	}
 }