+++ src/Resources/contao/languages/de/tl_mitgliederverwaltung_tournaments.php
@@ -5,7 +5,15 @@
  */
 $GLOBALS['TL_LANG']['tl_mitgliederverwaltung_tournaments']['tournament_legend'] = 'Turnier';
 $GLOBALS['TL_LANG']['tl_mitgliederverwaltung_tournaments']['titel'] = array('Titel', 'Geben Sie hier den Turniertitel ein');
-$GLOBALS['TL_LANG']['tl_mitgliederverwaltung_tournaments']['date'] = array('Startdatum', 'Startdatum im Format TT.MM.JJJJ');
+$GLOBALS['TL_LANG']['tl_mitgliederverwaltung_tournaments']['kennziffer'] = array('Kennzeichen', 'Kennzeichen bzw. Kennziffer des Turniers/der Turnierserie');
+$GLOBALS['TL_LANG']['tl_mitgliederverwaltung_tournaments']['registrationDate'] = array('Meldedatum', 'Meldedatum lt. Ausschreibung im Format TT.MM.JJJJ');
+$GLOBALS['TL_LANG']['tl_mitgliederverwaltung_tournaments']['startDate'] = array('Startdatum', 'Startdatum lt. Ausschreibung im Format TT.MM.JJJJ');
+$GLOBALS['TL_LANG']['tl_mitgliederverwaltung_tournaments']['art'] = array('Art', 'Art des Turniers');
+$GLOBALS['TL_LANG']['tl_mitgliederverwaltung_tournaments']['nenngeld'] = array('Nenngeld', 'Nenngeld in Euro');
+
+$GLOBALS['TL_LANG']['tl_mitgliederverwaltung_tournaments']['turnierleiter_legend'] = 'Turnierleiter';
+$GLOBALS['TL_LANG']['tl_mitgliederverwaltung_tournaments']['turnierleiterName'] = array('Name', 'Name des Turnierleiters');
+$GLOBALS['TL_LANG']['tl_mitgliederverwaltung_tournaments']['turnierleiterEmail'] = array('E-Mail', 'E-Mail-Adresse des Turnierleiters');
 
 $GLOBALS['TL_LANG']['tl_mitgliederverwaltung_tournaments']['applications_legend'] = 'Bewerbungen';
 $GLOBALS['TL_LANG']['tl_mitgliederverwaltung_tournaments']['name'] = array('Spieler', 'Name des Spielers');
@@ -39,3 +47,9 @@ $GLOBALS['TL_LANG']['tl_mitgliederverwaltung_tournaments']['members'] = array('M
 
 $GLOBALS['TL_LANG']['tl_mitgliederverwaltung_tournaments']['bewerbungen'] = array('Bewerbungen', 'Anzahl der Bewerbungen');
 $GLOBALS['TL_LANG']['tl_mitgliederverwaltung_tournaments']['zusagen'] = array('Status', 'Anzahl der Zu- oder Absagen');
+
+$GLOBALS['TL_LANG']['tl_mitgliederverwaltung_tournaments']['art_options'] = array
+(
+	'n' => 'Nationales Turnier',
+	'i' => 'Internationales Turnier',
+);